Moved generator and solver to separate packages
